make CashAddress a base class and use a new BitcoinCashAddress subclass
This commit makes `CashAddress` a base class and creates a subclass `BitcoinCashAddress` that is now used throughout the codebase instead of `CashAddress`.
This will allow in the following commits to make `CashAddress` support any prefix (`cashHRP`) rather than just `"bitcoincash"`, so that it can be re-used for an `ECashAddress` subclass.